feat: create EventSubtypePicker component
