+++ agir/front/components/mockData/eventSubtypes.json
@@ -1,0 +1,173 @@
+[
+  {
+    "id": 6,
+    "label": "diffusion-tracts",
+    "description": "Diffusion de tracts",
+    "color": "#c2306c",
+    "icon": null,
+    "iconName": "exclamation",
+    "type": "A"
+  },
+  {
+    "id": 7,
+    "label": "reunion-groupe",
+    "description": "Réunion régulière du groupe",
+    "color": "#4a64ac",
+    "icon": null,
+    "iconName": "comments",
+    "type": "G"
+  },
+  {
+    "id": 9,
+    "label": "porte-a-porte",
+    "description": "Porte-à-porte",
+    "color": "#c2306c",
+    "icon": null,
+    "iconName": "exclamation",
+    "type": "A"
+  },
+  {
+    "id": 12,
+    "label": "collage",
+    "description": "Collage d'affiches",
+    "color": "#C9462C",
+    "icon": null,
+    "iconName": "paint-brush",
+    "type": "A"
+  },
+  {
+    "id": 57,
+    "label": "action-solidarite",
+    "description": "Action de solidarité",
+    "color": "#C9462C",
+    "icon": null,
+    "iconName": "handshake-o",
+    "type": "A"
+  },
+  {
+    "id": 4,
+    "label": "caravane",
+    "description": "Caravane d'accès aux droits",
+    "color": "#c2306c",
+    "icon": null,
+    "iconName": "truck",
+    "type": "A"
+  },
+  {
+    "id": 5,
+    "label": "inscription-listes",
+    "description": "Point d'inscriptions sur les listes électorales",
+    "color": "#0098B6",
+    "icon": null,
+    "iconName": "edit",
+    "type": "A"
+  },
+  {
+    "id": 11,
+    "label": "réunion publique",
+    "description": "Réunion publique",
+    "color": "#e14b35",
+    "icon": null,
+    "iconName": "bullhorn",
+    "type": "M"
+  },
+  {
+    "id": 13,
+    "label": "autre-evenement",
+    "description": "Autre",
+    "color": "#49b37d",
+    "icon": null,
+    "iconName": "calendar",
+    "type": "O"
+  },
+  {
+    "id": 14,
+    "label": "covoiturage",
+    "description": "Covoiturage",
+    "color": "#49b37d",
+    "icon": null,
+    "iconName": "car",
+    "type": "O"
+  },
+  {
+    "id": 16,
+    "label": "soutien",
+    "description": "Soutien à une manifestation, un rassemblement",
+    "color": "#c9462c",
+    "icon": null,
+    "iconName": "calendar",
+    "type": "O"
+  },
+  {
+    "id": 22,
+    "label": "soiree-accueil",
+    "description": "Soirée d'accueil des nouveaux membres",
+    "color": "#4a64ac",
+    "icon": null,
+    "iconName": "calendar",
+    "type": "G"
+  },
+  {
+    "id": 31,
+    "label": "ecoute-collective",
+    "description": "Ecoute collective",
+    "color": "#0098B6",
+    "icon": null,
+    "iconName": "desktop",
+    "type": "M"
+  },
+  {
+    "id": 37,
+    "label": "diffusion-programme",
+    "description": "Diffusion de notre programme",
+    "color": "#00acee",
+    "icon": null,
+    "iconName": "book",
+    "type": "A"
+  },
+  {
+    "id": 41,
+    "label": "pique-nique",
+    "description": "Pique-nique, apéro citoyen",
+    "color": "#C9462C",
+    "icon": null,
+    "iconName": "cutlery",
+    "type": "M"
+  },
+  {
+    "id": 51,
+    "label": "fête-locale",
+    "description": "Fête locale",
+    "color": "#0098B6",
+    "icon": null,
+    "iconName": "star-half-o",
+    "type": "M"
+  },
+  {
+    "id": 52,
+    "label": "formations",
+    "description": "Formations",
+    "color": "#3b5998",
+    "icon": null,
+    "iconName": "book",
+    "type": "O"
+  },
+  {
+    "id": 54,
+    "label": "projection-debat",
+    "description": "Projection-débat",
+    "color": "#00b400",
+    "icon": null,
+    "iconName": "film",
+    "type": "M"
+  },
+  {
+    "id": 56,
+    "label": "atelier-pancartes",
+    "description": "Atelier de fabrication de pancartes",
+    "color": "#0098B6",
+    "icon": null,
+    "iconName": "paint-brush",
+    "type": "O"
+  }
+]
